MSI-1601: Fix failed static tests on 2.3-develop branch
