+++ app/code/Magento/InventoryCatalogAdminUi/Controller/Adminhtml/BulkAbstract.php
@@ -1,0 +1,85 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Magento\InventoryCatalogAdminUi\Controller\Adminhtml;
+
+use Magento\Backend\App\Action;
+use Magento\Catalog\Model\ResourceModel\Product\CollectionFactory;
+use Magento\Framework\Controller\ResultFactory;
+use Magento\InventoryCatalogAdminUi\Model\BulkSessionProductsStorage;
+use Magento\Ui\Component\MassAction\Filter;
+
+abstract class BulkAbstract extends Action
+{
+    /**
+     * @see _isAllowed()
+     */
+    const ADMIN_RESOURCE = 'Magento_Catalog::products';
+
+    /**
+     * @var Filter
+     */
+    private $filter;
+
+    /**
+     * @var BulkSessionProductsStorage
+     */
+    private $bulkSessionProductsStorage;
+
+    /**
+     * @var CollectionFactory
+     */
+    private $collectionFactory;
+
+    /**
+     * @param Action\Context $context
+     * @param CollectionFactory $collectionFactory
+     * @param Filter $filter
+     * @param BulkSessionProductsStorage $bulkSessionProductsStorage
+     * @SuppressWarnings(PHPMD.LongVariable)
+     */
+    public function __construct(
+        Action\Context $context,
+        CollectionFactory $collectionFactory,
+        Filter $filter,
+        BulkSessionProductsStorage $bulkSessionProductsStorage
+    ) {
+        parent::__construct($context);
+
+        $this->collectionFactory = $collectionFactory;
+        $this->filter = $filter;
+        $this->bulkSessionProductsStorage = $bulkSessionProductsStorage;
+    }
+
+    /**
+     * @return \Magento\Framework\App\ResponseInterface|\Magento\Framework\Controller\ResultInterface
+     */
+    public function execute()
+    {
+        try {
+            $collection = $this->filter->getCollection($this->collectionFactory->create());
+        } catch (\Exception $e) {
+            $this->messageManager->addErrorMessage($e->getMessage());
+            $redirect = $this->resultFactory->create(ResultFactory::TYPE_REDIRECT);
+            return $redirect->setPath('catalog/product/index');
+        }
+
+        $this->bulkSessionProductsStorage->setProductsSkus($collection->getColumnValues('sku'));
+
+        $resultPage = $this->resultFactory->create(ResultFactory::TYPE_PAGE);
+        $resultPage->getConfig()->getTitle()->prepend(__($this->getTitle()));
+
+        return $resultPage;
+    }
+
+    /**
+     * Return title for page
+     *
+     * @return string
+     */
+    abstract protected function getTitle(): string;
+}
