+++ app/code/Magento/InventoryCatalogAdminUi/Controller/Adminhtml/Inventory/BulkTransfer.php
@@ -7,72 +7,15 @@ declare(strict_types=1);
 
 namespace Magento\InventoryCatalogAdminUi\Controller\Adminhtml\Inventory;
 
-use Magento\Backend\App\Action;
-use Magento\Catalog\Model\ResourceModel\Product\CollectionFactory;
-use Magento\Framework\Controller\ResultFactory;
-use Magento\InventoryCatalogAdminUi\Model\BulkSessionProductsStorage;
-use Magento\Ui\Component\MassAction\Filter;
+use Magento\InventoryCatalogAdminUi\Controller\Adminhtml\BulkAbstract;
 
-class BulkTransfer extends Action
+class BulkTransfer extends BulkAbstract
 {
     /**
-     * @see _isAllowed()
+     * @inheritdoc
      */
-    const ADMIN_RESOURCE = 'Magento_Catalog::products';
-
-    /**
-     * @var Filter
-     */
-    private $filter;
-
-    /**
-     * @var BulkSessionProductsStorage
-     */
-    private $bulkSessionProductsStorage;
-
-    /**
-     * @var CollectionFactory
-     */
-    private $collectionFactory;
-
-    /**
-     * @param Action\Context $context
-     * @param CollectionFactory $collectionFactory
-     * @param Filter $filter
-     * @param BulkSessionProductsStorage $bulkSessionProductsStorage
-     * @SuppressWarnings(PHPMD.LongVariable)
-     */
-    public function __construct(
-        Action\Context $context,
-        CollectionFactory $collectionFactory,
-        Filter $filter,
-        BulkSessionProductsStorage $bulkSessionProductsStorage
-    ) {
-        parent::__construct($context);
-
-        $this->filter = $filter;
-        $this->bulkSessionProductsStorage = $bulkSessionProductsStorage;
-        $this->collectionFactory = $collectionFactory;
-    }
-
-    /**
-     * @return \Magento\Framework\App\ResponseInterface|\Magento\Framework\Controller\ResultInterface
-     */
-    public function execute()
+    protected function getTitle(): string
     {
-        try {
-            $collection = $this->filter->getCollection($this->collectionFactory->create());
-        } catch (\Exception $e) {
-            $this->messageManager->addErrorMessage($e->getMessage());
-            $redirect = $this->resultFactory->create(ResultFactory::TYPE_REDIRECT);
-            return $redirect->setPath('catalog/product/index');
-        }
-
-        $this->bulkSessionProductsStorage->setProductsSkus($collection->getColumnValues('sku'));
-
-        $resultPage = $this->resultFactory->create(ResultFactory::TYPE_PAGE);
-        $resultPage->getConfig()->getTitle()->prepend(__('Bulk Inventory Transfer'));
-
-        return $resultPage;
+        return 'Bulk Inventory Transfer';
     }
 }
