+++ app/code/Magento/InventoryCatalogAdminUi/view/adminhtml/templates/inventory_transfer.phtml
@@ -121,7 +121,7 @@ $sources = $viewModel->getSources();
     </form>
 </div>
 
-<script type="text/javascript">
+<script>
     require(['jquery'], function ($) {
         $('#back').click(function () {
             self.location.href='<?= /* @noEscape */ $block->getUrl('catalog/product/index') ?>';
