+++ app/code/Magento/InventoryCatalogAdminUi/view/adminhtml/templates/source_selection.phtml
@@ -97,7 +97,7 @@ $sources = $viewModel->getSources();
     </form>
 </div>
 
-<script type="text/javascript">
+<script>
     require(['jquery'], function ($) {
         $('#back').click(function () {
             self.location.href='<?= /* @noEscape */ $block->getUrl('catalog/product/index') ?>';
